refactor(portfolio): rename NoNeuronsCard to NoProjectsCard (#6162)
# Motivation

The Portfolio page displays information about Tokens and Projects.

Initially, the component intended for users who have not staked anything
in any project was called `NoNeuronsCard`. However, this is inaccurate;
it should be named `NoProjectsCard`.

# Changes

- Renames `NoNeuronsCard` to `NoProjectsCard`

# Tests

- Tests should pass as they did before

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary
