+++ frontend/src/lib/components/portfolio/NoProjectsCard.svelte
